+++ examples/grass_mesh.rs
@@ -43,9 +43,9 @@ fn setup_grass(
     };
     commands.insert_resource(store);
 
-    let height_map = asset_server.load("grass_height_map.png");
+    let y_map_image = asset_server.load("grass_height_map.png");
 
-    let height_map = HeightMap { height_map };
+    let y_map = YMap { y_map: y_map_image };
     let density_map = asset_server.load("grass_density_map.png");
 
     let density_map = DensityMap {
@@ -56,7 +56,7 @@ fn setup_grass(
     commands.spawn(WarblersBundle {
         grass_mesh,
         density_map,
-        height_map,
+        y_map,
         aabb: Aabb::from_min_max(Vec3::ZERO, Vec3::new(100., 10., 100.)),
         ..default()
     });
@@ -69,7 +69,7 @@ struct GrassMeshStore {
     pub default_handle: Handle<Mesh>,
 }
 
-// Swapps the mesh type if TAB is pressed
+// Swap the mesh type if TAB is pressed
 fn swap_grass_mesh(
     mut commands: Commands,
     mut queue: Query<(Entity, &mut Handle<Mesh>)>,
