+++ src/density_map.rs
@@ -1,25 +1,24 @@
 //! Contains the implementation of the [`DensityMap`] component
 use bevy::{asset::Handle, ecs::component::Component, reflect::Reflect, render::texture::Image};
-/// The density map defining the density of grass at specific positions.
-/// White pixels corresponds to dense areas.
+
+/// The density map defines the density of grass blades at a given positions.
 ///
-/// Usually, this component is used in the [`WarblersBundle`](crate::bundle::WarblersBundle)
-///
-/// The area covered by the density map is defined by the area of the [`Aabb`](bevy::render::primitives::Aabb) component
-///
+/// The area covered by the density map is defined by the area of the [`Aabb`](bevy::render::primitives::Aabb) component.
 /// The [`DensityMap`] texture will be scaled over the complete area
 /// Often a small density map is enough to cover big areas!
 ///
-/// For a simple example, take a look at the `load_grass` example.
+/// For a simple example, take a look at the [`load_grass`](https://github.com/emiongit/warbler_grass/latest/example/load_grass.rs) example.
 #[derive(Reflect, Clone, Component)]
 pub struct DensityMap {
     /// The density map of the grass chunk.
     ///
-    /// Should be a gray scale image
+    /// Should be ideally gray scale image for memory efficency.
+    /// White pixels corresponds to dense areas.
+    /// Black pixels correspond to sparse areas.
     pub density_map: Handle<Image>,
     /// The density of the grass.
     ///
-    /// If the density is high, more grass is spawned in a dense area
+    /// If the density is high, more grass is spawned in a dense area.
     pub density: f32,
 }
 /// A density map can be created from the image alone
